requirements.txt
Flask
mysql-connector-python
pandas
secure-smtplib
py-emails
scikit-learn
seaborn
pillow